refactor: packaged request
