+++ api/src/main/java/com/thesurvey/api/controller/UserController.java
@@ -1,7 +1,7 @@
 package com.thesurvey.api.controller;
 
 import com.thesurvey.api.domain.User;
-import com.thesurvey.api.dto.UserRegisterRequestDto;
+import com.thesurvey.api.dto.request.UserRegisterRequestDto;
 import com.thesurvey.api.dto.UserInfoDto;
 import com.thesurvey.api.exception.ErrorMessage;
 import com.thesurvey.api.exception.ExceptionMapper;
